backend/ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
google-generativeai==0.3.1
httpx==0.27.0
python-dotenv==1.0.1
redis==5.0.1
tenacity==8.2.3
